Refactor on Gemini suggestions from PR Review #9
