+++ src/RepairCafe.Api/Modules/ModuleRegistry.cs
@@ -1,28 +1,23 @@
+using System.Collections.Concurrent;
 using System.Reflection;
 
 namespace RepairCafe.Api.Modules;
 
 public static class ModuleRegistry
 {
-    private static readonly List<Assembly> _domainAssemblies = [];
-    private static readonly List<Assembly> _applicationAssemblies = [];
+    private static readonly ConcurrentDictionary<Assembly, byte> _domainAssemblies = new();
+    private static readonly ConcurrentDictionary<Assembly, byte> _applicationAssemblies = new();
 
-    public static IEnumerable<Assembly> DomainAssemblies => _domainAssemblies.AsReadOnly();
-    public static IEnumerable<Assembly> ApplicationAssemblies => _applicationAssemblies.AsReadOnly();
+    public static IEnumerable<Assembly> DomainAssemblies => _domainAssemblies.Keys;
+    public static IEnumerable<Assembly> ApplicationAssemblies => _applicationAssemblies.Keys;
 
     public static void RegisterDomainAssembly(Assembly assembly)
     {
-        if (!_domainAssemblies.Contains(assembly))
-        {
-            _domainAssemblies.Add(assembly);
-        }
+        _domainAssemblies.TryAdd(assembly, 0);
     }
 
     public static void RegisterApplicationAssembly(Assembly assembly)
     {
-        if (!_applicationAssemblies.Contains(assembly))
-        {
-            _applicationAssemblies.Add(assembly);
-        }
+        _applicationAssemblies.TryAdd(assembly, 0);
     }
 }
