+++ src/Shared/RepairCafe.Shared.Infrastructure/DomainEvents/DomainEventDispatcher.cs
@@ -1,6 +1,4 @@
-﻿using Microsoft.EntityFrameworkCore;
-using Newtonsoft.Json;
-using RepairCafe.Shared.Infrastructure.Outbox;
+﻿using Newtonsoft.Json;
 using RepairCafe.Shared.Infrastructure.Outbox.Models;
 using RepairCafe.Shared.Infrastructure.Persistence;
 using RepairCafe.Shared.Kernel.Abstractions;
